# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AlphaTrade Pro API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ForgotPasswordRequest {
  phone: string;
}

export interface ResetPasswordRequest {
  token: string;
  newPassword: string;
}

export type RegisterDeviceRequestPlatform = typeof RegisterDeviceRequestPlatform[keyof typeof RegisterDeviceRequestPlatform];


export const RegisterDeviceRequestPlatform = {
  ios: 'ios',
  android: 'android',
} as const;

export interface RegisterDeviceRequest {
  token: string;
  platform: RegisterDeviceRequestPlatform;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface MessageResponse {
  message: string;
}

export type TradeType = typeof TradeType[keyof typeof TradeType];


export const TradeType = {
  intraday: 'intraday',
  swing: 'swing',
  positional: 'positional',
} as const;

export type RiskLevel = typeof RiskLevel[keyof typeof RiskLevel];


export const RiskLevel = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export type RecommendationStatus = typeof RecommendationStatus[keyof typeof RecommendationStatus];


export const RecommendationStatus = {
  active: 'active',
  target_hit: 'target_hit',
  stop_loss_hit: 'stop_loss_hit',
  hold: 'hold',
  partial_profit: 'partial_profit',
  closed: 'closed',
} as const;

export type SubscriptionType = typeof SubscriptionType[keyof typeof SubscriptionType];


export const SubscriptionType = {
  trial: 'trial',
  premium: 'premium',
  expired: 'expired',
} as const;

export interface User {
  id: number;
  name: string;
  phone: string;
  email?: string;
  subscriptionType: SubscriptionType;
  trialStartDate?: string | null;
  trialExpiryDate?: string | null;
  premiumExpiryDate?: string | null;
  isBlocked: boolean;
  isAdmin: boolean;
  createdAt: string;
}

export type AdminUser = User & ({
  totalPayments?: number;
  lastPaymentDate?: string | null;
});

export interface RegisterRequest {
  name: string;
  phone: string;
  email?: string;
  password: string;
}

export interface LoginRequest {
  phone: string;
  password: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Recommendation {
  id: number;
  stockName: string;
  nseSymbol: string;
  buyPrice: number;
  targetPrice: number;
  stopLoss: number;
  tradeType: TradeType;
  riskLevel: RiskLevel;
  notes?: string | null;
  status: RecommendationStatus;
  exitPrice?: number | null;
  pnlPercent?: number | null;
  pnlAbsolute?: number | null;
  date: string;
  createdAt: string;
  updatedAt: string;
}

export interface RecommendationList {
  data: Recommendation[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface CreateRecommendationRequest {
  stockName: string;
  nseSymbol: string;
  buyPrice: number;
  targetPrice: number;
  stopLoss: number;
  tradeType: TradeType;
  riskLevel: RiskLevel;
  notes?: string;
  date?: string;
}

export interface UpdateRecommendationRequest {
  stockName?: string;
  nseSymbol?: string;
  buyPrice?: number;
  targetPrice?: number;
  stopLoss?: number;
  tradeType?: TradeType;
  riskLevel?: RiskLevel;
  notes?: string;
  date?: string;
}

export interface UpdatePnlRequest {
  exitPrice: number;
  status: RecommendationStatus;
  notes?: string;
}

export interface DashboardStats {
  todayRecommendationsCount: number;
  monthlyProfitPercent: number;
  totalTrades: number;
  winRate: number;
  subscriptionType: SubscriptionType;
  trialDaysRemaining?: number | null;
  premiumDaysRemaining?: number | null;
  bestTrade?: Recommendation | null;
  worstTrade?: Recommendation | null;
  recentActivity?: Recommendation[];
}

export interface MonthlyPerformance {
  month: string;
  totalTrades: number;
  wins: number;
  losses: number;
  winRate: number;
  totalPnlPercent: number;
  bestTradePercent: number;
  worstTradePercent: number;
}

export interface MonthlyReport {
  month: string;
  summary: MonthlyPerformance;
  recommendations: Recommendation[];
}

export interface AdminStats {
  totalUsers: number;
  activeTrialUsers: number;
  premiumUsers: number;
  expiredUsers: number;
  totalRevenue: number;
  monthlyRevenue: number;
  totalRecommendations: number;
  averageWinRate: number;
  newUsersThisMonth: number;
  newUsersLastMonth: number;
}

export interface AdminUserList {
  data: AdminUser[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface AdminUpdateUserRequest {
  isBlocked?: boolean;
  subscriptionType?: SubscriptionType;
  premiumExpiryDate?: string;
}

export type RevenueAnalyticsMonthlyItem = {
  month: string;
  revenue: number;
  newSubscriptions: number;
  renewals: number;
};

export interface RevenueAnalytics {
  monthly: RevenueAnalyticsMonthlyItem[];
  totalRevenue: number;
  avgMonthlyRevenue: number;
}

export type UserGrowthAnalyticsMonthlyItem = {
  month: string;
  newUsers: number;
  totalUsers: number;
  premiumConversions: number;
};

export interface UserGrowthAnalytics {
  monthly: UserGrowthAnalyticsMonthlyItem[];
}

export interface SubscriptionStatus {
  subscriptionType: SubscriptionType;
  trialStartDate?: string | null;
  trialExpiryDate?: string | null;
  premiumExpiryDate?: string | null;
  daysRemaining?: number | null;
  isActive: boolean;
}

export interface PaymentOrder {
  orderId: string;
  amount: number;
  currency: string;
  keyId: string;
}

export interface VerifyPaymentRequest {
  razorpayOrderId: string;
  razorpayPaymentId: string;
  razorpaySignature: string;
}

export type PaymentStatus = typeof PaymentStatus[keyof typeof PaymentStatus];


export const PaymentStatus = {
  pending: 'pending',
  success: 'success',
  failed: 'failed',
} as const;

export interface Payment {
  id: number;
  amount: number;
  currency: string;
  status: PaymentStatus;
  razorpayOrderId?: string | null;
  razorpayPaymentId?: string | null;
  createdAt: string;
}

export type SendNotificationRequestTargetType = typeof SendNotificationRequestTargetType[keyof typeof SendNotificationRequestTargetType];


export const SendNotificationRequestTargetType = {
  all: 'all',
  premium: 'premium',
  trial: 'trial',
  specific: 'specific',
} as const;

export interface SendNotificationRequest {
  title: string;
  body: string;
  targetType?: SendNotificationRequestTargetType;
  userIds?: number[];
}

export interface Notification {
  id: number;
  title: string;
  body: string;
  targetType: string;
  sentCount: number;
  createdAt: string;
}

export type ListRecommendationsParams = {
/**
 * Filter by date (YYYY-MM-DD)
 */
date?: string;
/**
 * Filter by month (YYYY-MM)
 */
month?: string;
tradeType?: TradeType;
status?: RecommendationStatus;
page?: number;
limit?: number;
};

export type AdminListUsersParams = {
page?: number;
limit?: number;
search?: string;
status?: AdminListUsersStatus;
};

export type AdminListUsersStatus = typeof AdminListUsersStatus[keyof typeof AdminListUsersStatus];


export const AdminListUsersStatus = {
  all: 'all',
  trial: 'trial',
  premium: 'premium',
  expired: 'expired',
} as const;