# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AlphaTrade Pro API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with phone/email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "200":
          description: Reset initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── RECOMMENDATIONS ────────────────────────────────────────────────────────
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [all, trial, premium, expired]
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserList"
  /admin/users/{id}:
    get:
      operationId: adminGetUser
      tags: [admin]
      summary: Get user details (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
    put:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Update user (admin) - block, extend subscription
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
...
  /notifications/register-device:
    post:
      operationId: registerDevice
      tags: [notifications]
      summary: Register device push token
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDeviceRequest"
      responses:
        "200":
          description: Token registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /notifications/send:
    post:
      operationId: sendNotification
      tags: [notifications]
      summary: Send push notification (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendNotificationRequest"
      responses:
        "200":
          description: Notification sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /notifications/history:
    get:
      operationId: getNotificationHistory
      tags: [notifications]
      summary: Get notification history (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notification history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"

# ─── COMPONENTS ──────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ForgotPasswordRequest:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
    ResetPasswordRequest:
      type: object
      required: [token, newPassword]
      properties:
        token:
          type: string
        newPassword:
          type: string
    RegisterDeviceRequest:
      type: object
      required: [token, platform]
      properties:
        token:
          type: string
        platform:
          type: string
          enum: [ios, android]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    SubscriptionType:
      type: string
      enum: [trial, premium, expired]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        subscriptionType:
          $ref: "#/components/schemas/SubscriptionType"
        trialStartDate:
          type: string
          format: date-time
          nullable: true
        trialExpiryDate:
          type: string
          format: date-time
          nullable: true
        premiumExpiryDate:
          type: string
          format: date-time
          nullable: true
        isBlocked:
          type: boolean
        isAdmin:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, phone, subscriptionType, isBlocked, isAdmin, createdAt]
    AdminUser:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          properties:
            totalPayments:
              type: number
            lastPaymentDate:
              type: string
              format: date-time
              nullable: true
    RegisterRequest:
      type: object
      required: [name, phone, password]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        password:
          type: string
    LoginRequest:
      type: object
      required: [phone, password]
      properties:
        phone:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
...
    AdminUserList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [data, total, page, limit, totalPages]
    AdminUpdateUserRequest:
      type: object
      properties:
        isBlocked:
          type: boolean
        subscriptionType:
          $ref: "#/components/schemas/SubscriptionType"
        premiumExpiryDate:
          type: string
          format: date-time
...
    SendNotificationRequest:
      type: object
      required: [title, body]
      properties:
        title:
          type: string
        body:
          type: string
        targetType:
          type: string
          enum: [all, premium, trial, specific]
          default: all
        userIds:
          type: array
          items:
            type: integer
    Notification:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        body:
          type: string
        targetType:
          type: string
        sentCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, title, body, targetType, sentCount, createdAt]