# lib/api-zod/src/generated/types/registerDeviceRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AlphaTrade Pro API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RegisterDeviceRequestPlatform } from './registerDeviceRequestPlatform';

export interface RegisterDeviceRequest {
  token: string;
  platform: RegisterDeviceRequestPlatform;
}