# lib/api-zod/src/generated/types/sendNotificationRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AlphaTrade Pro API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SendNotificationRequestTargetType } from './sendNotificationRequestTargetType';

export interface SendNotificationRequest {
  title: string;
  body: string;
  targetType?: SendNotificationRequestTargetType;
  userIds?: number[];
}