# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AlphaTrade Pro API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SubscriptionType } from './subscriptionType';

export interface User {
  id: number;
  name: string;
  phone: string;
  email?: string;
  subscriptionType: SubscriptionType;
  trialStartDate?: Date | null;
  trialExpiryDate?: Date | null;
  premiumExpiryDate?: Date | null;
  isBlocked: boolean;
  isAdmin: boolean;
  createdAt: Date;
}